Fix missing wound display on floater HP bar
Add in missing controller parts to update the Wounds value.

The UI for the floater HP bar already has the code to update the UI, so you can see the changes immediately there when the buttons are clicked.  The focus info panel HP bar does not reflect wounds.  This change doesn't update that.